Time: 31 ms (50.04%) | Memory: 36.6 MB (49.82%) - LeetSync

+++ 518-coin-change-ii/coin-change-ii.cpp
@@ -17,7 +17,7 @@ public:
                     take = dp[i][j-coins[i]];
                 }
 
-                dp[i][j] = (int)(skip + (long long)(take));
+                dp[i][j] = (long long)skip + take;
             }
         }
